Remove executable bit from non-executable files
Git tracks whether or not a file is executable, so that it can set the
appropriate file mode on Unix-like systems. Windows does not track a
file's executable-ness in the same way, and so git sometimes sets the
wrong file mode on Windows. Remove the executable bit from the various
files in this repo that are not scripts or executable binaries.