Restore unhandled keyword from pre_range
The call 'arg_keyword= is_keyword( trailing_keywords );'
in line 1034 may have consumed/removed a keyword
from 'trailing_keywords', but it hasn't been handled so far.
Therefore restore the previous value of 'trailing_keywords',
so that the keyword can be handled further below.

This e.g. fixes the command

    pdftk A=some_file.pdf cat Aeast output rotated.pdf

so that rotation is actually applied.

Fixes #11