Specify target Java version to build for
When not specifying the target Java version, the version of the
installed JDK is targetted. This leads to errors when trying to execute
the generated JAR file on a system with an earlier Java version than the
one it was built with. Specify the lowest possible target version
according to the README to ensure that the compiled JAR runs on as many
systems as possible.